Exporting RowMapper classes
