+++ src/main/java/com/gmail/alexander/realestatestore/repository/BuyerDAO.java
@@ -1,6 +1,7 @@
 package com.gmail.alexander.realestatestore.repository;
 
 import com.gmail.alexander.realestatestore.models.costumers.Buyer;
+import com.gmail.alexander.realestatestore.repository.rowmappers.BuyerRowMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -25,37 +26,12 @@ public class BuyerDAO {
     @Autowired
     private EmployeeDAO employeeDAO;
 
-    /**
-     * This class is used to Map Data to The objects fields.
-     */
-    class BuyerRowMapper implements RowMapper<Buyer> {
 
-        /**
-         * Maps data to Object
-         *
-         * @param rs     is used to store all data from database and then extracted to the object`s fields.
-         * @param rowNum Store what number of a row are we in.
-         * @return Object from given class.
-         * @throws SQLException
-         */
-        @Override
-
-        public Buyer mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Buyer buyer = new Buyer();
-            buyer.setId(rs.getInt("id"));
-            buyer.setName(rs.getString("name"));
-            buyer.setPhone(rs.getString("phone"));
-            buyer.setBudget(rs.getDouble("budget"));
-            buyer.getEmployee().setId(rs.getInt("employee_id"));
-            return buyer;
-        }
-    }
     /**
      * Finds all Records of Buyer
      *
      * @return
      */
-
     public List<Buyer> findAll() {
         return jdbcTemplate.query("SELECT Buyer.id, budget, name, phone, employee_id FROM Buyer INNER JOIN Customer ON Buyer.ID=Customer.ID ", new BuyerRowMapper());
 
@@ -63,15 +39,16 @@ public class BuyerDAO {
 
     /**
      * Gets a list of buyers by Employee`s id.
+     *
      * @param id given employee`s id
      * @return
      */
     public List<Buyer> findBuyersByEmployeeId(int id) {
-        return jdbcTemplate.query("SELECT Buyer.id, budget, name, phone, employee_id FROM Buyer INNER JOIN Customer ON Buyer.ID=Customer.ID WHERE Buyer.employee_id=?",  new Object[]{id}, new BuyerRowMapper());
+        return jdbcTemplate.query("SELECT Buyer.id, budget, name, phone, employee_id FROM Buyer INNER JOIN Customer ON Buyer.ID=Customer.ID WHERE Buyer.employee_id=?", new Object[]{id}, new BuyerRowMapper());
     }
 
-    public Buyer findTopByOrderByIdDesc(){
-        return jdbcTemplate.queryForObject("SELECT Buyer.id, budget, name, phone, employee_id FROM Buyer INNER JOIN Customer ON Buyer.ID=Customer.ID ORDER BY ID DESC LIMIT 1",  new BuyerRowMapper());
+    public Buyer findTopByOrderByIdDesc() {
+        return jdbcTemplate.queryForObject("SELECT Buyer.id, budget, name, phone, employee_id FROM Buyer INNER JOIN Customer ON Buyer.ID=Customer.ID ORDER BY ID DESC LIMIT 1", new BuyerRowMapper());
     }
 
     /**
@@ -90,29 +67,28 @@ public class BuyerDAO {
      * @param id for deletion
      * @return
      */
-
     public int deleteById(int id) {
         return jdbcTemplate.update("DELETE FROM Buyer WHERE id=?", id);
     }
+
     /**
      * This is used to insert record to the database.
+     *
      * @param buyer
      * @return
      */
-
     public int insert(Buyer buyer) {
 
-        return jdbcTemplate.update("INSERT INTO Buyer (id, BUDGET, EMPLOYEE_ID) " + "VALUES((SELECT id FROM CUSTOMER WHERE phone=?), ? , (SELECT id FROM EMPLOYEE WHERE id=? ))",
-               buyer.getPhone(), buyer.getBudget(), employeeDAO.randomEmployee());
+        return jdbcTemplate.update("INSERT INTO Buyer (id, BUDGET, EMPLOYEE_ID) " + "VALUES((SELECT id FROM CUSTOMER WHERE phone=?), ? , (SELECT id FROM EMPLOYEE WHERE id=? ))", buyer.getPhone(), buyer.getBudget(), employeeDAO.randomEmployee());
     }
 
     /**
      * Updates record from database by id.
+     *
      * @param buyer updated version of the record.
      * @return confirmation code.
      */
     public int update(Buyer buyer) {
-        return jdbcTemplate.update("UPDATE Buyer " + "SET BUDGET=?  " + "WHERE id=(SELECT id FROM CUSTOMER WHERE id=?    )",
-                buyer.getBudget(), buyer.getId());
+        return jdbcTemplate.update("UPDATE Buyer " + "SET BUDGET=?  " + "WHERE id=(SELECT id FROM CUSTOMER WHERE id=?    )", buyer.getBudget(), buyer.getId());
     }
 }
