+++ src/main/java/com/gmail/alexander/realestatestore/repository/LandDAO.java
@@ -2,6 +2,7 @@ package com.gmail.alexander.realestatestore.repository;
 
 import com.gmail.alexander.realestatestore.models.concrete.Agency;
 import com.gmail.alexander.realestatestore.models.realesates.Land;
+import com.gmail.alexander.realestatestore.repository.rowmappers.LandRowMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -20,40 +21,8 @@ import java.util.List;
  */
 @Repository
 public class LandDAO {
-
-
     @Autowired
     private JdbcTemplate jdbcTemplate;
-
-    /**
-     * This class is used to Map Data to The objects fields.
-     */
-    class LandRowMapper implements RowMapper<Land> {
-        /**
-         * Maps data to Object
-         *
-         * @param rs     is used to store all data from database and then extracted to the object`s fields.
-         * @param rowNum Store what number of a row are we in.
-         * @return Object from given class.
-         * @throws SQLException
-         */
-        @Override
-        public Land mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Land land = new Land();
-            land.setId(rs.getInt("id"));
-            land.setAddress(rs.getString("address"));
-            land.setDescription(rs.getString("description"));
-            land.setPrice(rs.getDouble("price"));
-            land.setSizeOfRealEstate(rs.getInt("size_of_real_estate"));
-            land.setRealEstateType(rs.getString("real_estate_type"));
-            land.setLandType(rs.getString("land_type"));
-            land.setRegulated(rs.getBoolean("is_regulated"));
-            land.setSold(rs.getBoolean("sold"));
-            land.getSeller().setId(rs.getInt("seller_id"));
-            land.getEmployee().setId(rs.getInt("employee_id"));
-            return land;
-        }
-    }
 
     /**
      * Finds all Records of Agency
@@ -76,7 +45,7 @@ public class LandDAO {
         return jdbcTemplate.queryForObject("SELECT Land.id, address, price, description, SIZE_OF_REAL_ESTATE, REAL_ESTATE_TYPE, land_type, is_sold, is_regulated, seller_id, employee_id FROM Land INNER JOIN Property ON Land.ID=Property.ID WHERE Property.ID=?", new Object[]{id}, new BeanPropertyRowMapper<Land>(Land.class));
     }
 
-    public Agency findTopByOrderByIdDesc(){
+    public Agency findTopByOrderByIdDesc() {
         return jdbcTemplate.queryForObject("SELECT Land.id, address, price, description, SIZE_OF_REAL_ESTATE, REAL_ESTATE_TYPE, land_type, is_sold, is_regulated, seller_id, employee_id FROM Land INNER JOIN Property ON Land.ID=Property.ID ORDER BY ID DESC LIMIT 1", new BeanPropertyRowMapper<Agency>(Agency.class));
     }
 
@@ -93,6 +62,7 @@ public class LandDAO {
 
     /**
      * This is used to insert record to the database.
+     *
      * @param land
      * @return
      */
