+++ src/main/java/com/gmail/alexander/realestatestore/repository/MainRepository.java
@@ -1,0 +1,25 @@
+package com.gmail.alexander.realestatestore.repository;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class MainRepository {
+    @Autowired
+    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  ApartmentDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+//    @Autowired
+//    private  AgencyDAO agencyDAO;
+
+}
