+++ src/main/java/com/gmail/alexander/realestatestore/repository/rowmappers/EmployeeRowMapper.java
@@ -1,0 +1,30 @@
+package com.gmail.alexander.realestatestore.repository.rowmappers;
+
+import com.gmail.alexander.realestatestore.models.concrete.Employee;
+import org.springframework.jdbc.core.RowMapper;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+/**
+ * This class is used to Map Data to The objects fields.
+ */
+public class EmployeeRowMapper implements RowMapper<Employee> {
+    /**
+     * Maps data to Object
+     *
+     * @param rs     is used to store all data from database and then extracted to the object`s fields.
+     * @param rowNum Store what number of a row are we in.
+     * @return Object from given class.
+     * @throws SQLException
+     */
+    @Override
+    public Employee mapRow(ResultSet rs, int rowNum) throws SQLException {
+        Employee employee = new Employee();
+        employee.setId(rs.getInt("id"));
+        employee.setName(rs.getString("name"));
+        employee.setPhone(rs.getString("phone"));
+        employee.getAgency().setId(rs.getInt("agency_id"));
+        return employee;
+    }
+}
